Update to Debug Protocol 1.36.x:
- Introduces a capability completionTriggerCharacters for announcing
the characters that a frontend UI should use to trigger completion.
- Adds an optional sortText attribute to the CompletionItem type.
- In the schema change “number” to “integer” in cases where a float
makes no sense, e.g. for IDs.
- Fixed type error in EvaluateResponse.memoryReference - now a String
instead of Long to match protocol.

+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/DataBreakpointInfoArguments.java
@@ -23,10 +23,11 @@ import org.eclipse.xtext.xbase.lib.util.ToStringBuilder;
 public class DataBreakpointInfoArguments {
   /**
    * Reference to the Variable container if the data breakpoint is requested for a child of the container.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long variablesReference;
+  private Integer variablesReference;
   
   /**
    * The name of the Variable's child to obtain data breakpoint information for. If variableReference isn't provided,
@@ -37,20 +38,22 @@ public class DataBreakpointInfoArguments {
   
   /**
    * Reference to the Variable container if the data breakpoint is requested for a child of the container.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getVariablesReference() {
+  public Integer getVariablesReference() {
     return this.variablesReference;
   }
   
   /**
    * Reference to the Variable container if the data breakpoint is requested for a child of the container.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setVariablesReference(final Long variablesReference) {
+  public void setVariablesReference(final Integer variablesReference) {
     this.variablesReference = variablesReference;
   }
