+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/SetVariableResponse.java
@@ -36,29 +36,31 @@ public class SetVariableResponse {
   
   /**
    * If variablesReference is > 0, the new value is structured and its children can be retrieved by passing
-   * variablesReference to the VariablesRequest.
+   * variablesReference to the VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long variablesReference;
+  private Integer variablesReference;
   
   /**
    * The number of named child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long namedVariables;
+  private Integer namedVariables;
   
   /**
    * The number of indexed child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long indexedVariables;
+  private Integer indexedVariables;
   
   /**
    * The new value of the variable.
@@ -97,22 +99,22 @@ public class SetVariableResponse {
   
   /**
    * If variablesReference is > 0, the new value is structured and its children can be retrieved by passing
-   * variablesReference to the VariablesRequest.
+   * variablesReference to the VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getVariablesReference() {
+  public Integer getVariablesReference() {
     return this.variablesReference;
   }
   
   /**
    * If variablesReference is > 0, the new value is structured and its children can be retrieved by passing
-   * variablesReference to the VariablesRequest.
+   * variablesReference to the VariablesRequest. The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setVariablesReference(final Long variablesReference) {
+  public void setVariablesReference(final Integer variablesReference) {
     this.variablesReference = variablesReference;
   }
   
@@ -120,11 +122,12 @@ public class SetVariableResponse {
    * The number of named child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getNamedVariables() {
+  public Integer getNamedVariables() {
     return this.namedVariables;
   }
   
@@ -132,10 +135,11 @@ public class SetVariableResponse {
    * The number of named child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setNamedVariables(final Long namedVariables) {
+  public void setNamedVariables(final Integer namedVariables) {
     this.namedVariables = namedVariables;
   }
   
@@ -143,11 +147,12 @@ public class SetVariableResponse {
    * The number of indexed child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getIndexedVariables() {
+  public Integer getIndexedVariables() {
     return this.indexedVariables;
   }
   
@@ -155,10 +160,11 @@ public class SetVariableResponse {
    * The number of indexed child variables.
    * <p>
    * The client can use this optional information to present the variables in a paged UI and fetch them in chunks.
+   * The value should be less than or equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setIndexedVariables(final Long indexedVariables) {
+  public void setIndexedVariables(final Integer indexedVariables) {
     this.indexedVariables = indexedVariables;
   }
